+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Refactor removal of arguments for formatters (from custom class to Linter generic class)
 - Perl: lint files with no extension containing Perl shebang
 - Add automerge for PR issues from linter versions updates
+- Fix ignored root files issue
 
 ## [4.4.0] 2020-11-05